Add files via upload
Select a Dataset for Assignment through Official Website of Central Statistics Office. Link has been Mentioned below

Link : https://data.cso.ie/table/SES01

After That we create a file Csv File on Microsoft Excel and make a Dataset of Mean and median earnings per hour and paid weekly hours. 

After that we create a Coding file through Jupiter Notebook along with extension of IPYNB and create a Libraries for data importing.